requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
asyncpg==0.30.0
psycopg2-binary==2.9.9
alembic==1.13.0
python-dotenv==1.0.1
pydantic==2.9.0
pydantic-settings==2.5.0
httpx==0.27.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
groq==0.11.0
google-cloud-vision==3.7.0
google-auth==2.35.0
gspread==6.1.0
Pillow==10.4.0
apscheduler==3.10.4
slowapi==0.1.9
numpy==1.26.4
psutil==5.9.8
